# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics Inventory API — multi-tenant inventory SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthAck" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Email not verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
      required: [error]
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    AuthAck:
      type: object
      properties:
        ok: { type: boolean }
        message: { type: string }
      required: [ok, message]
    AuthSession:
      type: object
      properties:
        ok: { type: boolean }
        user:
          oneOf:
            - { $ref: "#/components/schemas/User" }
            - type: "null"
    SignupBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: ["string", "null"] }
      required: [email, password]
    LoginBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    TokenBody:
      type: object
      properties:
        token: { type: string }
      required: [token]
    EmailBody:
      type: object
      properties:
        email: { type: string }
      required: [email]
    ResetPasswordBody:
      type: object
      properties:
        token: { type: string }
        password: { type: string }
      required: [token, password]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
...
    User:
      type: object
      properties:
        id: { type: integer }
        emailVerified: { type: boolean }
        email: { type: string }
        name: { type: ["string", "null"] }
        isSuperAdmin: { type: boolean }
      required: [id, email, name, isSuperAdmin, emailVerified]
...